requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bitsandbytes==0.42.0
bm25s==0.2.5
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
filelock==3.16.1
filetype==1.2.0
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.4
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
kiwisolver==1.4.7
llama-cloud==0.1.6
llama-index==0.12.2
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.2
llama-index-embeddings-gemini==0.3.0
llama-index-embeddings-huggingface==0.5.2
llama-index-embeddings-jinaai==0.4.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.2
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-index-retrievers-bm25==0.5.0
llama-parse==0.5.15
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
openai==1.56.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pypdf==5.1.0
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.17.0
yarl==1.18.3
